+++ src/code/math/nth_term_lrec.hpp
@@ -11,7 +11,7 @@ CEXP mint nth_term_lrec(u64 n, vec<mint> CR a, vec<mint> CR bm) {
   if (n < a.size()) return a[n];
   assert(!bm.empty() && bm[0] == 1);
   poly q(bm);
-  return q.strip(), bostan_mori(n, (poly(a) * q).pre(q.size() - 1), q);
+  return q.strip(), bostan_mori(n, (poly(a) * q).pre((u32)q.size() - 1), q);
 }
 template <class poly, std::same_as<TPN poly::val_t> mint>
 CEXP mint nth_term_lrec(u64 n, vec<mint> CR a) {
