+++ src/test_cpverifier/library-checker-graph/enumerate_triangles.test.cpp
@@ -15,7 +15,7 @@ int main() {
   vecptu edges(m);
   std::cin >> edges;
   u64 ans = 0;
-  auto func = [&](u32 u, u32 v, u32 w) { ans = (ans + (u128)x[u] * x[v] * x[w]) % 998244353; };
+  auto func = [&](u32 u, u32 v, u32 w) { ans = u32((ans + (u128)x[u] * x[v] * x[w]) % 998244353); };
   tifa_libs::graph::ringenum3(n, edges, func);
   std::cout << ans << '\n';
   return 0;
